whelp/requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.4.1
celery==4.4.1
redis==3.4.1
pdfminer3==2018.12.3.0
